% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pns.R
\name{fastpns}
\alias{fastpns}
\title{Computes fast PNS decomposition}
\usage{
fastpns(
  x,
  n.pc = "Full",
  sphere.type = "seq.test",
  mean.type = "Frechet",
  alpha = 0.1,
  R = 100,
  nlast.small.sphere = 1,
  output = TRUE
)
}
\arguments{
\item{x}{data on sphere}

\item{n.pc}{number of pc's (default 'Full')}

\item{sphere.type}{type of sphere decompistion (default 'seq.test')}

\item{mean.type}{type of mean (default 'Frechet')}

\item{alpha}{significant level for test (default .1)}

\item{R}{the number of bootstrap samples to be evaluated for the sequential test. The default is 100.}

\item{nlast.small.sphere}{(default 1)}

\item{output}{show output (default \code{TRUE})}
}
\value{
EuclidData euclidean representation
}
\description{
Computes fast PNS decomposition
}
