% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pns.R
\name{fastPNSe2s}
\alias{fastPNSe2s}
\title{Computes fast PNS euclidean to sphere}
\usage{
fastPNSe2s(res, pns)
}
\arguments{
\item{res}{PNS embedding}

\item{pns}{list of PNS structure}
}
\value{
approx sphere representation
}
\description{
Computes fast PNS euclidean to sphere
}
